Deleted useless files & advertlist render in index
